Adjust resources files.
