+++ app/src/main/java/tw/jizah/popocast/ui/theme/Dimens.kt
@@ -1,0 +1,19 @@
+package tw.jizah.popocast.ui.theme
+
+import androidx.compose.ui.unit.dp
+
+object Dimens {
+    val m1 = 4.dp
+    val m2 = 8.dp
+    val m3 = 12.dp
+    val m4 = 16.dp
+    val m5 = 20.dp
+    val m6 = 24.dp
+    val m7 = 28.dp
+    val m8 = 32.dp
+    val m9 = 36.dp
+    val m10 = 40.dp
+
+    val playerControllerSize = 80.dp
+    val playerControllerTextPadding = m1
+}
