--- app/src/main/java/tw/jizah/popocast/ui/theme/Resource.kt
@@ -1,21 +1,0 @@
-package tw.jizah.popocast.ui.theme
-
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.unit.dp
-
-object Colors {
-    val playerControllerBg = gray900
-    val playerControllerTitleText = Color.White
-    val playerControllerOwnerText = gray500
-    val playerControllerProgress = Color.Yellow
-}
-
-object Dimens {
-    val playerControllerSize = 80.dp
-    val playerControllerTextPadding = 4.dp
-}
-
-object Drawables {
-    const val iconPlay = android.R.drawable.ic_media_play
-    const val iconPause = android.R.drawable.ic_media_pause
-}
