+++ app/src/main/java/tw/jizah/popocast/ui/theme/Theme.kt
@@ -5,26 +5,30 @@ import androidx.compose.material.MaterialTheme
 import androidx.compose.material.darkColors
 import androidx.compose.material.lightColors
 import androidx.compose.runtime.Composable
+import tw.jizah.popocast.ui.theme.Colors.purple200
+import tw.jizah.popocast.ui.theme.Colors.purple500
+import tw.jizah.popocast.ui.theme.Colors.purple700
+import tw.jizah.popocast.ui.theme.Colors.teal200
 
 private val DarkColorPalette = darkColors(
-        primary = purple200,
-        primaryVariant = purple700,
-        secondary = teal200
+    primary = purple200,
+    primaryVariant = purple700,
+    secondary = teal200
 )
 
 private val LightColorPalette = lightColors(
-        primary = purple500,
-        primaryVariant = purple700,
-        secondary = teal200
+    primary = purple500,
+    primaryVariant = purple700,
+    secondary = teal200
 
-        /* Other default colors to override
-    background = Color.White,
-    surface = Color.White,
-    onPrimary = Color.White,
-    onSecondary = Color.Black,
-    onBackground = Color.Black,
-    onSurface = Color.Black,
-    */
+    /* Other default colors to override
+background = Color.White,
+surface = Color.White,
+onPrimary = Color.White,
+onSecondary = Color.Black,
+onBackground = Color.Black,
+onSurface = Color.Black,
+*/
 )
 
 @Composable
@@ -36,9 +40,9 @@ fun PoPoCastTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composab
     }
 
     MaterialTheme(
-            colors = colors,
-            typography = PoPoCastTypography,
-            shapes = shapes,
-            content = content
+        colors = colors,
+        typography = PoPoCastTypography,
+        shapes = shapes,
+        content = content
     )
 }
